pyproject.toml
[tool.poetry]
name = "der-koerper-bot"
version = "0.1.0"
description = ""
authors = ["Gregor Weichbrodt <gregorweichbrodt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mastodon-py = "^1.8.1"
pydantic = "^2.7.1"
tenacity = "^8.2.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
